extracted dropdown row
